perf(wallet): report active wallets p3a when balances are fetched
