+++ components/brave_wallet_ui/common/slices/api.slice.ts
@@ -304,7 +304,6 @@ export const {
   useRemoveAccountMutation,
   useRemoveSitePermissionMutation,
   useRemoveUserTokenMutation,
-  useReportActiveWalletsToP3AMutation,
   useReportOnboardingActionMutation,
   useRequestSitePermissionMutation,
   useRestoreNetworksMutation,
